+++ sdk/keyvault/keyvault-admin/test/public/backupClient.abort.spec.ts
@@ -7,7 +7,7 @@ import { AbortController } from "@azure/abort-controller";
 import { KeyVaultBackupClient } from "../../src";
 import { authenticate } from "../utils/authentication";
 import { testPollerProperties } from "../utils/recorder";
-import { assertThrowsAbortError, getFolderName, getSasToken } from "../utils/common";
+import { assertThrowsAbortError, getSasToken } from "../utils/common";
 
 describe("Aborting KeyVaultBackupClient's requests", () => {
   let client: KeyVaultBackupClient;
@@ -46,13 +46,11 @@ describe("Aborting KeyVaultBackupClient's requests", () => {
 
   it("can abort beginRestore", async function() {
     const backupURI = `${blobStorageUri}/${generateFakeUUID()}`;
-    const folderName = getFolderName(backupURI);
-
     const controller = new AbortController();
     controller.abort();
 
     await assertThrowsAbortError(async () => {
-      await client.beginRestore(blobStorageUri, blobSasToken, folderName, {
+      await client.beginRestore(backupURI, blobSasToken, {
         ...testPollerProperties,
         abortSignal: controller.signal
       });
@@ -61,13 +59,12 @@ describe("Aborting KeyVaultBackupClient's requests", () => {
 
   it("can abort beginSelectiveRestore", async function() {
     const backupURI = `${blobStorageUri}/${generateFakeUUID()}`;
-    const folderName = getFolderName(backupURI);
 
     const controller = new AbortController();
     controller.abort();
 
     await assertThrowsAbortError(async () => {
-      await client.beginSelectiveRestore(blobStorageUri, blobSasToken, folderName, "Key Name", {
+      await client.beginSelectiveRestore("Key Name", backupURI, blobSasToken, {
         ...testPollerProperties,
         abortSignal: controller.signal
       });